+++ CHANGELOG.md
@@ -39,6 +39,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   Git HEAD to the parent of the fresh working-copy commit.
   [#9936](https://github.com/jj-vcs/jj/issues/9936)
 
+* Fixed crash in `jj log` involving hidden revisions and the
+  `log-graph-prioritize` revset.
+  [#9975](https://github.com/jj-vcs/jj/issues/9975)
+
 ## [0.44.0] - 2026-08-05
 
 ### Release highlights